# docs/openapi.yaml
openapi: "3.0.3"
...
info:
  title: Tiny Win API
  version: "1.0.0"
  description: |
    RESTful API for **Tiny Win** — the minimalist daily-win social platform.

    ### Core behaviours covered by this spec
    - **Auth** — register, login, logout, token refresh
    - **Wins** — create today's win (1 per user per day), retrieve own win, check status
    - **Feed** — friend feed (server-enforced lock until user has posted today)
    - **Reactions** — add / change / remove a reaction (🔥 👀 🤝)
    - **Streaks & Calendar** — user streak counters and win calendar

    ### Design principles
    - Server is source-of-truth for date boundaries (uses user's IANA timezone)
    - All write endpoints require `Idempotency-Key` header to prevent double-submit
    - Feed endpoint returns **403 FeedLocked** when user has not posted today
    - No public like/follower counts (anti-Instagram, PRD §2.3)

  contact:
    name: Tiny Win Engineering
    email: eng@tinywin.app
  license:
    name: Proprietary
servers:
  - url: https://api.tinywin.app/v1
    description: Production
  - url: https://staging-api.tinywin.app/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development

# ---------------------------------------------------------------------------
# Security Schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /auth/login` or `POST /auth/refresh`.
        Include in header: `Authorization: Bearer <token>`

  # -------------------------------------------------------------------------
  # Reusable Parameters
  # -------------------------------------------------------------------------
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID v4. Server stores this key and returns the cached
        response on duplicate requests within the same calendar day.
        Prevents double-submit (PRD §8, US-A3).
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
    WinId:
      name: win_id
      in: path
      required: true
      description: Unique identifier of the Tiny Win post.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateWinRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 120
          description: |
            Win text. 1–120 Unicode grapheme clusters after trim.
            Empty or whitespace-only content is rejected (PRD §5, Rules #3-4).
          example: "Hoàn thành 20 phút thiền định buổi sáng 🧘"
    Win:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c3d4e5f6-a1b2-7890-abcd-ef1234567890"
        author_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        author_username:
          type: string
          example: "hoanganh"
        author_display_name:
          type: string
          example: "Hoàng Anh"
        content:
          type: string
          example: "Hoàn thành 20 phút thiền định buổi sáng 🧘"
        date_key:
          type: string
          format: date
          description: Calendar date in user's timezone (server-computed)
          example: "2024-07-15"
        created_at:
          type: string
          format: date-time
          example: "2024-07-15T07:32:00Z"
        my_reaction:
          $ref: '#/components/schemas/ReactionType'
          nullable: true
          description: The authenticated user's current reaction on this win, or null.
    TodayWinStatus:
      type: object
      description: Tells the client whether the current user has posted today.
      properties:
        has_posted_today:
          type: boolean
          example: true
        date_key:
          type: string
          format: date
          example: "2024-07-15"
        win:
          $ref: '#/components/schemas/Win'
          nullable: true
          description: The win object if posted, null otherwise.

    # ── Reactions ───────────────────────────────────────────────────────────
    ReactionType:
      type: string
      enum: ["🔥", "👀", "🤝"]
      description: |
        Exactly three reaction types are supported (PRD §4.1, Rule #7):
        - 🔥 Fire
        - 👀 Eyes
        - 🤝 Handshake
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "WIN_ALREADY_EXISTS"
            message:
              type: string
              description: Human-readable explanation (calm, specific language per PRD §7)
              example: "Bạn đã đăng Tiny Win hôm nay rồi. Hẹn gặp lại ngày mai!"
            detail:
              type: string
              nullable: true
              description: Optional additional context
              example: "date_key=2024-07-15, existing_win_id=c3d4..."

  # -------------------------------------------------------------------------
  # Reusable Responses
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Token không hợp lệ hoặc đã hết hạn. Vui lòng đăng nhập lại."
...
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Không tìm thấy tài nguyên yêu cầu."
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Quá nhiều yêu cầu. Vui lòng thử lại sau."
...
security:
  - BearerAuth: []

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /wins:
    post:
      tags: [Wins]
      summary: Create today's Tiny Win
      description: |
        Posts the authenticated user's Tiny Win for today.

        **Business rules enforced server-side:**
        - 1 win per user per calendar day (user's IANA timezone, PRD §5 Rule #2)
        - Content: 1–120 grapheme clusters after trim (PRD §5 Rules #3-5)
        - Idempotency: duplicate `Idempotency-Key` returns the existing win (HTTP 200)
        - On success, the user's feed is unlocked for the rest of the day

        **Concurrent duplicate protection:** UNIQUE constraint on `(author_id, date_key)`
        combined with `Idempotency-Key` header prevents double-submit.
      operationId: createWin
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWinRequest'
            example:
              content: "Uống đủ 2L nước hôm nay 💧"
      responses:
        "201":
          description: Win created successfully. Feed is now unlocked.
          headers:
            Location:
              description: URL of the newly created win
              schema:
                type: string
                example: "/v1/wins/c3d4e5f6-a1b2-7890-abcd-ef1234567890"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Win'
        "200":
          description: |
            Idempotent response — this `Idempotency-Key` was already used today.
            Returns the existing win without creating a duplicate. (PRD US-A3, US-A4)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Win'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          description: |
            User has already posted a Tiny Win today under a **different** idempotency key.
            (PRD §5 Rule #2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "WIN_ALREADY_EXISTS"
                  message: "Bạn đã đăng Tiny Win hôm nay rồi. Hẹn gặp lại ngày mai!"
                  detail: "date_key=2024-07-15"
        "422":
          description: Content validation failed (blank, too long, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "CONTENT_INVALID"
                  message: "Nội dung phải có từ 1 đến 120 ký tự và không chỉ gồm khoảng trắng."
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /wins/today:
    get:
      tags: [Wins]
      summary: Get today's win status for the current user
      description: |
        Returns whether the current user has posted today and, if so, the win object.
        Used by the home screen to show/hide the post CTA (US-A1).
        Cached in Redis per user (TTL until end of day in user timezone).
      operationId: getTodayWinStatus
      responses:
        "200":
          description: Today's win status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayWinStatus'
              examples:
                posted:
                  summary: User has posted today
                  value:
                    has_posted_today: true
                    date_key: "2024-07-15"
                    win:
                      id: "c3d4e5f6-a1b2-7890-abcd-ef1234567890"
                      author_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                      content: "Uống đủ 2L nước hôm nay 💧"
                      date_key: "2024-07-15"
                      created_at: "2024-07-15T07:32:00Z"
                      my_reaction: null
                not_posted:
                  summary: User has not posted today
                  value:
                    has_posted_today: false
                    date_key: "2024-07-15"
                    win: null
        "401":
          $ref: '#/components/responses/Unauthorized'
  /wins/{win_id}:
    get:
      tags: [Wins]
      summary: Get a specific win by ID
      description: |
        Returns a win post. The requesting user must be the author or
        an accepted friend of the author (RLS enforced at DB layer).
      operationId: getWin
      parameters:
        - $ref: '#/components/parameters/WinId'
      responses:
        "200":
          description: Win found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Win'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Not authorized to view this win (not an accepted friend).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "FORBIDDEN"
                  message: "Bạn không có quyền xem bài viết này."
        "404":
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # FEED
  # ===========================================================================
...